+++ python/amici/__init__.py
@@ -23,8 +23,9 @@ models and turning them into C++ Python extensions.
 import importlib
 import os
 import re
+import sys
 from types import ModuleType as ModelModule
-from typing import Any, Dict, Optional, Union, Sequence, List
+from typing import Optional
 
 
 def _get_amici_path():
@@ -88,12 +89,6 @@ amiciModulePath = os.path.dirname(__file__)
 has_clibs = any([os.path.isfile(os.path.join(amici_path, wrapper))
                  for wrapper in ['amici.py', 'amici_without_hdf5.py']])
 
-AmiciModel = Union['amici.Model', 'amici.ModelPtr']
-AmiciSolver = Union['amici.Solver', 'amici.SolverPtr']
-AmiciExpData = Union['amici.ExpData', 'amici.ExpDataPtr']
-AmiciReturnData = Union['amici.ReturnData', 'amici.ReturnDataPtr']
-AmiciExpDataVector = Union['amici.ExpDataPtrVector', Sequence[AmiciExpData]]
-
 # Get version number from file
 with open(os.path.join(amici_path, 'version.txt')) as f:
     __version__ = f.read().strip()
@@ -128,65 +123,6 @@ if not _imported_from_setup():
         pass
 
 hdf5_enabled = 'readSolverSettingsFromHDF5' in dir()
-
-
-# Values are suffixes of `get[...]` and `set[...]` `amici.Model` methods.
-# If either the getter or setter is not named with this pattern, then the value
-# is a tuple where the first and second elements are the getter and setter
-# methods, respectively.
-model_instance_settings = [
-    'AddSigmaResiduals',
-    'AlwaysCheckFinite',
-    'FixedParameters',
-    'InitialStates',
-    'InitialStateSensitivities',
-    'MinimumSigmaResiduals',
-    ('nMaxEvent', 'setNMaxEvent'),
-    'Parameters',
-    'ParameterList',
-    'ParameterScale',  # getter returns a SWIG object
-    'ReinitializationStateIdxs',
-    'ReinitializeFixedParameterInitialStates',
-    'StateIsNonNegative',
-    'SteadyStateSensitivityMode',
-    ('t0', 'setT0'),
-    'Timepoints',
-]
-
-
-def get_model_settings(
-        model: AmiciModel,
-) -> Dict[str, Any]:
-    """Get model settings that are set independently of the compiled model.
-
-    :param model: The AMICI model instance.
-
-    :returns: Keys are AMICI model attributes, values are attribute values.
-    """
-    settings = {}
-    for setting in model_instance_settings:
-        getter = setting[0] if isinstance(setting, tuple) else f'get{setting}'
-        settings[setting] = getattr(model, getter)()
-        # TODO `amici.Model.getParameterScale` returns a SWIG object instead
-        # of a Python list/tuple.
-        if setting == 'ParameterScale':
-            settings[setting] = tuple(settings[setting])
-    return settings
-
-
-def set_model_settings(
-        model: AmiciModel,
-        settings: Dict[str, Any],
-) -> None:
-    """Set model settings.
-
-    :param model: The AMICI model instance.
-    :param settings: Keys are callable attributes (setters) of an AMICI model,
-        values are provided to the setters.
-    """
-    for setting, value in settings.items():
-        setter = setting[1] if isinstance(setting, tuple) else f'set{setting}'
-        getattr(model, setter)(value)
 
 
 class add_path:
